Correcting errors from SADXLVL2 name change.
